Refactor index types to use u16 for improved capacity in enemy and pet management

+++ crates/engine_new/src/actor/enemy.rs
@@ -132,7 +132,7 @@ impl EnemyManager {
         let mut manager = Self::new();
         for i in 0..count {
             manager.enemies.push(Enemy::raid_boss(
-                TargetIdx(i as u8),
+                TargetIdx(i as u16),
                 format!("Boss {}", i + 1),
             ));
         }
